+++ src/Universalis.Application/Views/WorldUploadCountView.cs
@@ -2,7 +2,7 @@
 {
     public class WorldUploadCountView
     {
-        public uint Count { get; set; }
+        public double Count { get; set; }
 
         public double Proportion { get; set; }
     }
